stream json data into kafka
