+++ generator/Dockerfile
@@ -4,6 +4,6 @@ RUN apt-get update && git clone https://github.com/edenhill/librdkafka.git && \
 	cd librdkafka && ./configure --prefix /usr && make && make install && \
 	go get -u github.com/confluentinc/confluent-kafka-go/kafka
 
-ADD generator.go /opt/generator-app/generator.go
+ADD . /opt/generator-app/
 
-CMD go run /opt/generator-app/generator.go
+CMD go run /opt/generator-app/kafka-producer.go
