+++ generator/kafka-producer.go
@@ -1,0 +1,41 @@
+package main
+
+import (
+	"fmt"
+	"github.com/confluentinc/confluent-kafka-go/kafka"
+  "./generator"
+)
+
+func main() {
+
+	p, err := kafka.NewProducer(&kafka.ConfigMap{"bootstrap.servers": "kafka"})
+	if err != nil {
+		panic(err)
+	}
+
+	// Delivery report handler for produced messages
+	go func() {
+		for e := range p.Events() {
+			switch ev := e.(type) {
+			case *kafka.Message:
+				if ev.TopicPartition.Error != nil {
+					fmt.Printf("Delivery failed: %v\n", ev.TopicPartition)
+				} else {
+					fmt.Printf("Delivered message to %v\n", ev.TopicPartition)
+				}
+			}
+		}
+	}()
+
+	// Produce messages to topic (asynchronously)
+	topic := "myTopic"
+	for _, message := range generator.CreateEvents(100) {
+		p.Produce(&kafka.Message{
+			TopicPartition: kafka.TopicPartition{Topic: &topic, Partition: kafka.PartitionAny},
+			Value: []byte(message),
+		}, nil)
+	}
+
+	// Wait for message deliveries
+	p.Flush(15 * 1000)
+}
